feat: physical names in the entire Construct Library (#2894)
This PR changes the xyzName attribute of all Resources in the Construct Library to be of type PhysicalName,
and adds an AWS linter rule that checks ensures conformance.

The name of the property can be any ending substring of the base name of the resource with the "Name" suffix -
for example, if my resource is AWS::MyService::AwesomeFoobar, the name can be foobarName or awesomeFoobarName.

BREAKING CHANGE: all <resource>Name attributes had their type changed from string to cdk.PhysicalName.