refactor(core): use "string" for physical name props (#3011)
We realized there is no need to use a strong type (`PhysicalName`) for resource physical names. Eventually users want to be able to either use the default, pass an explicit string for the name or request that the framework generate a name for them, conditionally (or, in the future, unconditionally) to whether the resource is being referenced across environments.

To that end, all physical name props have been converted back to `string` and a special token marker `PhysicalName.GENERATE_IF_NEEDED` was introduced which can be used to indicate that physical name will be generated if needed for cross environment references.

This change also modifies the protected API provided the `Resource` class:
- `physicalName` returns the value that should be passed to the CFN resource
- `getResourceNameAttribute(nameAttr)` returns a token that should be exposed as the resource name (e.g. `bucketName`).
- `getResourceArnAttribute(arnAttr, components)` returns a token that should bexposed as the resource's ARN (e.g. `bucketArn`).

All of these are "environment-sensitive" and will return "the right" value, depending on the resource's poster and configuration.