refactor(core): rename @aws-cdk/cdk to @aws-cdk/core (#2932)
Rename the module @aws-cdk/cdk to @aws-cdk/core since this is a more suitable name for it. 

Language-specific names:
- Python: aws_cdk.core (renamed)
- .NET: Amazon.CDK (not renamed)
- Java: only artifactId changed to `core` but package and groupId are still `software.amazon.awscdk`

This change completely deletes the @aws-cdk/cdk module from the repo as a temporary measure, to ensure there are no lingering references. A subsequent commit will introduce @aws-cdk/cdk as a deprecated package.

Fixed #2733 

BREAKING CHANGE: The `@aws-cdk/cdk` module was renamed to @aws-cdk/core
* **Python:** The `aws_cdk.cdk` module was renamed to `aws_cdk.core`
* **Java:** The artifact `cdk` in groupId `software.amazon.awscdk` was renamed to `core`