refactor(codepipeline): introduce IAction and unify the Action.bind() signature (#3012)
This brings the Action `bind()` API in line with our conventions.

It also introduces an `IAction` interface for those who want to work with the low-level Action interface.

The Action class has been moved from the aws-codepipeline to the aws-codepipeline-actions module.

This API is much more flexible, and I show the capabilities by changing the implementation of the `PipelineDeployStackAction` from `@app-delivery`.

BREAKING CHANGE: `app-delivery.PipelineDeployStackAction` is now a `codepipeline.IAction` instead of a construct.