chore: Remove unused types (#2874)
Added a linter rule to prohibit declaration of unused interfaces, data
types and enums. Unused means it is no reachable from any class. Certain
packages can declare interfaces for standard APIs, but in the general
case, an un-reachable interface / data type / enum is not useful as
it cannot be used to interact with the declared constructs.