fix(l1): output "any" instead of "object" for CFN "Json" (#3057)
As a work around to a JSII bug (awslabs/jsii#555), stop emitting
"object" as the type for fields denoted as "Primitive: JSON" in the
CloudFormation resource specification, and instead emit an "any" type.

This will cause the JSII model to not have a "JSON" field, instead an
"any" field, which advertises correct behavior in de/serialization.