fix(cli): allow bootstrapping the default environment
The common case of having an app with a stack that's environment-agnostic (e.g. does not specify "env") will cause this stack
to synthesize a manifest targeted at aws://unknown-account/unknown-region. When "cdk bootstrap" is invoked with a stack like that
it will now bootstrap the default environment (the one configured in the CLI).