feat(codepipeline): generate a Role for every AWS-owned Action used in a Pipeline. (#3105)
Instead of granting permissions to the Pipeline's Role,
which results in a Policy with a very wide range of permissions,
generate a separate Role for each AWS-owned Action added to the Pipeline.
The Pipeline Role will assume each Action-specific Role
(permissions are granted for that automatically)
before the Action is executed.

Also allow overriding this default behavior by specifying a Role for AWS-owned Actions.