chore: remove trailing space in MD for stability banner (#3188)
The trailing space that was created when inserting an empty line in the
banner's message was very hsotile to manually editing the file with an
editor configured to prune trailing white space on save.

Added a `trimRight()` call to remove any trailing whitespace, so that
the banner never includes trailing whitespace, making the files nicer
to otherwise edit manually without risking to break pkglint compliance.