fix(logs): fix infinite retention for jsii users (#3250)
The retention period was still accepting the value `Infinity` for
"infinite retention", even though the property has been changed to an
enum instead of `number`.

Even though this apparently works for TypeScript customers, jsii
customers will not be able to pass in a number where an enum value
is expected, and so won't be able to configure infinite retention.

Add an `INFINITE` value to the enum.