fix(codepipeline): Lambda doesn't have permissions to the pipeline bucket
The Lambda invoke action was missing granting permissions to the artifact bucket,
which meant attempts to use the credentials passed to it in the event resulted in an "access denied"
when attempting to read and/or write the action pipeline artifacts.
The fix is to grant the action role permissions to the artifact bucket
(read-write if the action has any outputs, just read if the action has only inputs).

Fixes #3274