chore: transform private references to be local
In order for dependabot to be able to identify package updates, it needs
to be able to resolve all packages in the dependency closure. Private
packages are not published to the npmjs.com registry and fail resolving.

This changes the private packages to use local path references within
the mono-repo, so that dependabot can work normally. This can also
enable standard use of `npm audit`.