fix(codepipeline): invoked Lambda doesn't have permissions to the pipeline bucket (#3303)
The Lambda invoke action was missing granting permissions to the artifact bucket,
which meant attempts to use the credentials passed to it in the event resulted in an "access denied"
when trying to read and/or write the action pipeline artifacts.
The fix is to grant the action role permissions to the artifact bucket
(write if the action has any outputs, read if the action has any inputs).

Fixes #3274