chore: fix lockfile references to local packages
Realized that the paths in the lockfiles were not relative to the correct location
for the first-layer of transitive dependencies of private packages - those need
to be relative to the directory where the lockfile is located.

The actual implementation is a little tricky as it appears node 12 seems to be
re-homing those references when you require the package.json file, but node 8
does not.