fix(cli): correctly handle tags when deploying multiple stacks (#3455)
Fixes a bug where stacks being declared with different tags end up with
identical tags when deployed to CloudFormation, due to the first tagged stack's
tags overriding the value of `options.tags`.

This change makes it so that the `options.tags` value is not overridden, and the
actual tag list to be used is determined in the same way for each stack that
will be deployed.

Fixes #3471