feat(eks): default capacity (#3633)
* feat(eks): default vpc

When defining an EKS cluster, the `vpc` property is now
optional. If not specified, a VPC will be defined with
the default configuration for this environment.

It can then be accessed through `cluster.vpc`.

Fixes #3541

* feat(eks): default capacity

By default, allocate 2x m5.large instances when defining an EKS cluster. This can be configured
through `defaultCapacity` and `defaultCapacityInstance` or by setting `defaultCapacity` to `0`
and invoking `cluster.addCapacity`.

BREAKING CHANGE: clusters will be created with a default capacity of x2 m5.large instances. You can specify `defaultCapacity: 0` if you wish to disable.

* Fix typo

* fix a few merge errors