chore: convert empty arrays to undefined
The CloudFormation templates created by CDK should not include empty arrays if the property is not required. The templates should be as concise and clean as possible.

This mostly uses the option to ignore empty lists in the methods of Lazy. However, in some cases we simply check for an empty list inline.