fix(lambda): generate correct metrics for aliases
Lambda supports metrics on individual aliases by specifying the metric
dimensions with `FunctionName` set to the underlying function's `Name`,
and `ResourceName` set to the qualified function `Name` (i.e:
`FunctionName:aliasName`).

The previous iteration of this code used the function's ARN instead of
it's name, resulting in incorrect metrics.

Fixes #3724