feat(cloudfront): define lambda@edge as resolvable resource (#2861)
* feat(cloudfront): define lambda@edge as resolvable resource

This declaration is required for deploying custom resources as lamdba
association, which by itself is required to deploy a lambda@edge for a
stack which is in a different region as 'us-east-1'.

Relates to #1575

* test: simplified test case by just test resources required for the case

* feat: allow to create a version from an arn

This commit allows to reference a lambda from a different region and use
that as function association.

* refactor: resolve conflicts

* refactor: update based on review

* refactor: use version arn as function arn

* Fixing package.json

* Fix tests