fix(init-templates): remove dependency on @types/node
That was not required anymore since we no longer require passing the
command line arguments to the `app.run()` call. This dependency has the
potential of causing collisions with the ES libraries included by the
TypeScript compiler, and specifying a correct version can be a delicate
exercise, especially as those templates have to keep working in the
future.

Additionally, pinned the TypeScript compiler version to the current
minor, as recommended by the TypeScript maintainers (TypeScript does not
honor SemVer, but guarantees no breaking change between stable releases,
which they indicate by issuing a new major version).

Fixes #3839