fix(codepipeline): insufficient deploy cross-account CFN role S3 permissions
The deploymentRole used in the CloudFormation deployment CodePipeline actions
was not granted explicit permissions to read from the pipeline's bucket.
This meant the deployment failed in the cross-account case
for templates that needed deploy-time access to the bucket
(for example: templates including a Lambda function),
as the pipeline bucket did not trust the deployment role.

Fixes #3765