fix(certificatemanager): increase minimum validation total timeout (#3914)
* fix(certificatemanager): add minimum backoff

Add a minimum component to the backoff and retry timer

* fix(certificatemanager): stub out sleep

Provide stub function for sleep to allow long running sleep to
cause test to fail

* fix(certificatemanager): spy on sleep

spy on sleep function to determine total time attempted to sleep

* fix(certificatemanager): Failing test

Failing test demonstrating min sleep is as low as 4.6 seconds.

* fix(certificatemanager): increase attempts

Increase max attempts in validation, still under 10,
but such that minimum sleep duration of one minute is reached
(but not more than 3 minutes).

Can adjust tests if window is not suitable, but I find that at least
30 seconds is needed in my region