refactor(assets): remove content hash attribute (#4003)
as part of the ci/cd design, we wish to reduce the surface area
of cdk templates to a minimum. content hash can only be determined
after the asset has been "prepared" (docker build/zip directory).

if the need arises to reintroduce this, we can always query the
content hash through a custom resource.

asset content hash is currently not used anywhere in the framework
or the aws construct library and introduces unwanted coupling between
the toolchain and the framework/app.

this change initially just removes the API from IAsset (and S3/ECR Asset classes)
to ensure code does not take a dependency on it. in a subsequent change we will
also remove the content hash cloudformation parameter.

BREAKING CHANGE: assets no longer expose a property `contentHash`. Use `sourceHash`
as a good approximation. if you have a strong use case for content hashes, please
raise a github issue and we will figure out a solution.