feat: publish construct hierarchy with metadata to cloud assembly
Publish the CDK construct hierarchy with the construct id and path to
the cloud assembly directory, with a corresponding manifest entry.

Motivation and proposal documented in RFC - https://github.com/aws/aws-cdk/pull/4053.

This change focuses on integrating generating the metadata into the
cloud assembly as part of the the 'synthesize' lifecycle phase.
Subsequent changes will enrich the metadata file with additional
information from the construct tree (see RFC).