fix(codepipeline): work around CodeBuild's pipeline key bug
CodeBuild has a bug where they ignore the encryption key of the pipeline's artifact bucket,
instead always using the project's key
(the account's default S3 key if the project key has not been set).
This makes the CodeBuild actions unusable in a cross-account pipeline,
as subsequent actions will get an 'Access Denied' error when trying to download the incorrectly encrypted artifacts.

The fix is to always set the project's key to be the same as the pipeline key in the CodeBuild action.

Fixes #4033