feat(s3-deployment): allow specifying memory limit
when deploying large files, users may need to increase the resource handler's memory configuration.

note: since custom resource handlers are singletons, we need to provision a handler for each memory configuration defined in the app. we do this by simply adding a suffix to the uuid of the singleton resource that includes the memory limit.

fixes #4058