chore(cx-api): bump CLOUD_ASSEMBLY_VERSION to 1.10.0 (#4259)
Since we've added a new type of artifact to the cloud assembly, namely
CDK_TREE, the protocol version needs to be bumped. This will result in
CDK customers being notified to upgrade the CDK CLI to >= 1.10.0.

To prevent this from occurring in the future, the change also changes
the behaviour when an unknown artifact type is detected when reading the
cloud assembly manifest. This ensures that the CLI will be forwards
compatible to cloud assemblies when new artifact types are added to
them.

See https://github.com/aws/aws-cdk/pull/4194#discussion_r328599305