fix(cli): make new CLI work with old assembly versions (#4307)
In the latest release, we changed the CX protocol (adding a new artifact
type), requiring users of a newer CLI to upgrade their CDK framework
libraries. This check was originally written to improve safety.

This particular change was actually backwards compatible, so add the
requisite code to the CLI to recognize and fix that condition.

Fixes #4294.