fix(codebuild): validate if a CodePipeline action that is cross-account does not have outputs (#4171)
CodeBuild does not honor the key set on the project if the key is from a different account. That means a cross-account CodeBuild action effectively cannot have outputs (as they will be written with the default S3 key of the CodeBuild account, which the other actions won't have access to).

Add validation that throws an error if there is an attempt to add a cross-account CodeBuild action with outputs.

Fixes #4032