fix(s3-deployment): lambda "src" not included in published module
As of v1.12.0, we customized the calculation of the custom resource hash
to use the contents of the "src" directory instead of the bundle.zip itself
in order to keep it deterministic (the bundle is created during build by
pulling in external dependencies and includes compiled python files which
caused it have a different hash in each build).

However, the "lambda/src" directory was explicitly excluded from the
module's tarball (in .npmignore), so when this module was used in production,
it could not find "lambda/src" in order to calculate the hash.

This fixes #4404 by removing `lambda/src` from `.npmignore`.