fix: use fixed dependency versions between CDK packages
Make CDK packages depend upon each other by using fixed versions
("1.12.0") instead of caret versions ("^1.12.0"). This change makes it
so that customers can choose to use an older version of the framework,
even if a newer one is available.

Without this change, that would not be possible since the caret
dependency would transitively pull in newer versions of the CDK
libraries and the dependency tree would end up with multiple copies
of the CDK library, risking runtime errors and breaking TypeScript
compilation.

This change does not completely solve the versioning issue yet,
as after this change customers and potential 3rd party construct
libraries they use still need to agree on the version of CDK to
be used (both need to use caret versioning, ensuring they always
get the latest version, or both need to use fixed versioning.
When version specifiers are mixed, multiple copies of the CDK will
be pulled into the dependency tree as soon as a ^caret version
can resolve to a new CDK release).

However, what this change DOES accomplish is stopping the worst bleeding
for users that only rely on 1st party packages, and for users for whom
it is okay to upgrade in lockstep with their 3rd party package provider.

NOTE: I would have preferred peerDependencies to have stayed at caret
versions, since that is true to the intent of the declaration (at least
this version is required), and it is more correct for what I feel should
be the final solution to this problem. However, it is `jsii` that forces
the values in `dependencies` and `peerDependencies` to be exactly the same.
Since coordinating a change in jsii makes it take too long to get this out,
and since it doesn't actually matter what the value of
`peerDependencies` is anyway in the presence of regular `dependencies`,
I have left it like this.