fix(core): child stack assembly metadata is duplidated on parent
When synthesizes the metadata section of a stack in the cloud assembly
manifest, we traverse the tree below the stack and collect the metadata
associated with all nodes in this tree. If this tree has a child stack,
we don't need to collect the metadata from it's node, because they would
appear in the metadata of that child stack.

Metadata annotated on nested stack nodes will be recorded by the
first non-nested stack parent.

Specifically this bug caused an issue when the child stack used assets.
Currently assets are reported through assembly metadata on the stack,
and they would be reported both on the child and on the parent (although
the CloudFormation parameters of the assets would not appear on the parent).
This resulted in a CloudFormation error that reads:

    ValidationError: Parameter values specified for a template which does not require them.

Therefore this fixes #2900