fix(eks): invalid arn when mapping users to rbac (#4549)
The ARN of IAM users was rendered with a region, but IAM resources
are global. This means that IAM users could not be mapped to Kubernetes
RBAC.

This fix adds a `user.userArn` attribute which correctly
renders the user's ARN and then uses this in the EKS library.

This fixes #4545